适配compose
适配 compose

+++ app/src/main/java/com/db/SqliteActivity.java
@@ -4,8 +4,6 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import androidx.appcompat.app.AppCompatActivity;
-import butterknife.ButterKnife;
-import butterknife.OnClick;
 import com.ilifesmart.utils.PersistentMgr;
 import com.ilifesmart.weather.R;
 
@@ -17,29 +15,21 @@ public class SqliteActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_sqlite);
-        ButterKnife.bind(this);
     }
 
-    @OnClick({R.id.sqlite_add, R.id.sqlite_delete, R.id.sqlite_update, R.id.sqlite_query, R.id.sqlite_upgrade})
     public void onViewClicked(View view) {
-        switch (view.getId()) {
-            case R.id.sqlite_add:
-                DBManager.getInstance().add();
-                break;
-            case R.id.sqlite_delete:
-                DBManager.getInstance().delete();
-                break;
-            case R.id.sqlite_update:
-                DBManager.getInstance().update();
-                break;
-            case R.id.sqlite_query:
-                String result = DBManager.getInstance().select();
-                Log.d(TAG, "onViewClicked: Result " + result);
-                break;
-            case R.id.sqlite_upgrade:
-                int version = PersistentMgr.readKV("SQLITE_VERSION", 1);
-                PersistentMgr.putKV("SQLITE_VERSION", version+1);
-                break;
+        if (view.getId() == R.id.sqlite_add) {
+            DBManager.getInstance().add();
+        } else if (view.getId() == R.id.sqlite_delete) {
+            DBManager.getInstance().delete();
+        } else if (view.getId() == R.id.sqlite_update) {
+            DBManager.getInstance().update();
+        } else if (view.getId() == R.id.sqlite_query) {
+            String result = DBManager.getInstance().select();
+            Log.d(TAG, "onViewClicked: Result " + result);
+        } else if (view.getId() == R.id.sqlite_upgrade) {
+            int version = PersistentMgr.readKV("SQLITE_VERSION", 1);
+            PersistentMgr.putKV("SQLITE_VERSION", version+1);
         }
     }
 }
